refactor: move Container component in accordance with atomic design
